Refactor tokens config into multiple files which separate concerns.
